pwm finished! ?

+++ software/iob-pwm.c
@@ -8,6 +8,6 @@ void pwm_init(int base_address){
 }
 
 //Set values on outputs
-void pwm_set_period(uint8_t period){
+void pwm_set_period(uint16_t period){
   IOB_PWM_SET_SPER(period);
 }
