+++ software/iob-pwm.h
@@ -7,4 +7,4 @@
 //Set GPIO base address
 void pwm_init(int base_address);
 
-void pwm_set_period(uint8_t period);
+void pwm_set_period(uint16_t period);
